Fixed compiler warnings and coded a Makefile. Currently at the end of testing phase

+++ philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: ouel-bou <ouel-bou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/04 11:33:06 by ouel-bou          #+#    #+#             */
-/*   Updated: 2024/10/17 12:55:19 by ouel-bou         ###   ########.fr       */
+/*   Updated: 2024/10/17 13:52:27 by ouel-bou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,8 @@ typedef struct s_philo
 {
 	int			philo_id;
 	bool		full;
-	size_t		meals_eaten;
-	size_t		last_meal;
+	long		meals_eaten;
+	long		last_meal;
 	t_mutex		*first_fork;
 	t_mutex		*second_fork;
 	t_table		*table;
@@ -40,10 +40,10 @@ typedef struct s_philo
 
 typedef struct s_clock
 {
-	size_t		t_to_die;
-	size_t		t_to_eat;
-	size_t		t_to_sleep;
-	size_t		start_time;
+	long		t_to_die;
+	long		t_to_eat;
+	long		t_to_sleep;
+	long		start_time;
 }				t_clock;
 
 typedef struct s_table
@@ -52,7 +52,7 @@ typedef struct s_table
 	bool		dead_flag;
 	bool		finish_flag;
 	bool		start_flag;
-	size_t		meals_num;
+	long		meals_num;
 	t_philo		*philos;
 	t_mutex		*forks;
 	t_mutex		table;
@@ -79,9 +79,9 @@ typedef enum e_status
 	DEAD,
 }	t_status;
 
-size_t	ft_strlen(char *s);
+long	ft_strlen(char *s);
 bool	check_num(char *av);
-size_t	ft_atol(char *s);
+long	ft_atol(char *s);
 bool	check_clock(t_clock *c);
 t_table	*check_args_input(int ac, char **av);
 void	err_exit(int status, char *msg);
@@ -93,12 +93,12 @@ void	init_data(t_table **data);
 
 bool	get_bool(t_mutex *mtx, bool flag);
 void	set_bool(t_mutex *mtx, bool *flag, bool value);
-size_t	get_num(t_mutex *mtx, size_t *num);
-void	set_num(t_mutex *mtx, size_t *des, size_t value);
-size_t	get_time(void);
-void	print_status(int phid, t_status status, size_t start_time, 
+long	get_num(t_mutex *mtx, long *num);
+void	set_num(t_mutex *mtx, long *des, long value);
+long	get_time(void);
+void	print_status(int phid, t_status status, long start_time, 
 			t_mutex *mtx);
-void	psleep(size_t us);
+void	psleep(long us);
 
 void	*philo_routine(void *data);
