+++ philo/srcs/simulation_utils.c
@@ -6,15 +6,15 @@
 /*   By: ouel-bou <ouel-bou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/05 22:39:26 by ouel-bou          #+#    #+#             */
-/*   Updated: 2024/10/17 13:12:23 by ouel-bou         ###   ########.fr       */
+/*   Updated: 2024/10/17 13:52:27 by ouel-bou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/philo.h"
 
-void	print_status(int phid, t_status status, size_t start_time, t_mutex *mtx)
+void	print_status(int phid, t_status status, long start_time, t_mutex *mtx)
 {
-	size_t	current_time;
+	long	current_time;
 
 	pthread_mutex_lock(mtx);
 	current_time = get_time() - start_time;
@@ -34,9 +34,9 @@ void	print_status(int phid, t_status status, size_t start_time, t_mutex *mtx)
 	pthread_mutex_unlock(mtx);
 }
 
-void	psleep(size_t us)
+void	psleep(long us)
 {
-	size_t	current_time;
+	long	current_time;
 
 	current_time = get_time();
 	while (get_time() - current_time < us / 1e3)
